+++ src/oci/database/models/__init__.py
@@ -23,6 +23,7 @@ from .autonomous_database_console_token_details import AutonomousDatabaseConsole
 from .autonomous_database_summary import AutonomousDatabaseSummary
 from .autonomous_database_wallet import AutonomousDatabaseWallet
 from .autonomous_db_preview_version_summary import AutonomousDbPreviewVersionSummary
+from .autonomous_db_version_summary import AutonomousDbVersionSummary
 from .autonomous_exadata_infrastructure import AutonomousExadataInfrastructure
 from .autonomous_exadata_infrastructure_maintenance_window import AutonomousExadataInfrastructureMaintenanceWindow
 from .autonomous_exadata_infrastructure_shape_summary import AutonomousExadataInfrastructureShapeSummary
@@ -158,6 +159,7 @@ database_type_mapping = {
     "AutonomousDatabaseSummary": AutonomousDatabaseSummary,
     "AutonomousDatabaseWallet": AutonomousDatabaseWallet,
     "AutonomousDbPreviewVersionSummary": AutonomousDbPreviewVersionSummary,
+    "AutonomousDbVersionSummary": AutonomousDbVersionSummary,
     "AutonomousExadataInfrastructure": AutonomousExadataInfrastructure,
     "AutonomousExadataInfrastructureMaintenanceWindow": AutonomousExadataInfrastructureMaintenanceWindow,
     "AutonomousExadataInfrastructureShapeSummary": AutonomousExadataInfrastructureShapeSummary,