+++ src/oci/regions.py
@@ -17,7 +17,9 @@ REGIONS_SHORT_NAMES = {
     'syd': 'ap-sydney-1',
     'ltn': 'uk-gov-london-1',
     'kix': 'ap-osaka-1',
-    'mel': 'ap-melbourne-1'
+    'mel': 'ap-melbourne-1',
+    'ams': 'eu-amsterdam-1',
+    'jed': 'me-jeddah-1'
 }
 REGION_REALMS = {
     'ap-melbourne-1': 'oc1',
@@ -28,8 +30,10 @@ REGION_REALMS = {
     'ap-tokyo-1': 'oc1',
     'us-phoenix-1': 'oc1',
     'us-ashburn-1': 'oc1',
+    'eu-amsterdam-1': 'oc1',
     'eu-frankfurt-1': 'oc1',
     'eu-zurich-1': 'oc1',
+    'me-jeddah-1': 'oc1',
     'uk-london-1': 'oc1',
     'ca-toronto-1': 'oc1',
     'sa-saopaulo-1': 'oc1',
@@ -58,8 +62,10 @@ REGIONS = [
     "ap-tokyo-1",
     "us-phoenix-1",
     "us-ashburn-1",
+    "eu-amsterdam-1",
     "eu-frankfurt-1",
     "eu-zurich-1",
+    "me-jeddah-1",
     "uk-london-1",
     "ca-toronto-1",
     "us-langley-1",